Removed the Jinja template code, it may come back but it was terrible. Moved all of the custom Unsync code into the core.py file. Re-arranged the code so that all of the unsync custom commands are imported from unsync. Also added a bunch of click imports to __init__ so that everything can be imported from unsync. Updated the builtin commands to use the new @unsync.command with builtin context data passing. This all needs testing.
