+++ unsync/unsync/commands/utils/move_tables_to_memory.py
@@ -1,13 +1,9 @@
-import click
+import unsync
 import petl
-
-from unsync.lib.unsync_data import pass_data
-from unsync.lib.unsync_commands import unsync
 
 
 @unsync.command()
-@click.option('--source', '-s', required=True, multiple=True, help='Name of the source data table.')
-@pass_data
+@unsync.option('--source', '-s', required=True, multiple=True, help='Name of the source data table.')
 def move_tables_to_memory(data, source):
     """Create a petl.MemorySource and copy data from the given tables into it. Can greatly decrease processing time for some operations."""
     for s in source:
