+++ unsync/unsync/commands/utils/stats.py
@@ -1,15 +1,10 @@
-import click
-import petl
-from unsync.lib.unsync_data import pass_data
-from unsync.lib.unsync_commands import unsync
+import unsync
 
 
 @unsync.command()
-@click.option('--source', '-s', required=True, help='Source table to print data from.')
-@pass_data
+@unsync.option('--source', '-s', required=True, help='Source table to print data from.')
 def stats(data, source):
     """Print a text representation of the data table for a given KIND."""
     d = data.get(source)
-    click.secho('Row Count: {}'.format(d.nrows()), fg='green')
-    click.secho('Column Count: {}'.format(len(d[0])), fg='green')
-
+    unsync.secho('Row Count: {}'.format(d.nrows()), fg='green')
+    unsync.secho('Column Count: {}'.format(len(d[0])), fg='green')
