--- unsync/unsync/lib/__init__.py
@@ -1,4 +1,0 @@
-from __future__ import absolute_import
-
-from .unsync_data import UnsyncData
-from .unsync_commands import UnsyncCommands, UnsyncCommand, Unsync
